requirements.txt
# Django
Django==5.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1


# Real-time
channels==4.1.0
channels-redis==4.2.0
redis==5.0.1

# Image Processing
Pillow==10.2.0
python-magic==0.4.27

# File Storage
cloudinary==1.38.0

# CORS
django-cors-headers

# Environment Variables
python-dotenv==1.0.1

# ASGI Server
daphne==4.0.0

# Development Tools
black==24.3.0
flake8==7.0.0

# API Documentation

drf-yasg==1.21.7

# Security
pyjwt==2.8.0

# Utilities
python-dateutil==2.9.0.post0


django-environ==0.11.2


